+++ docs/scripts/views/radio.vue
@@ -1,8 +1,8 @@
 <template>
   <div class="docs-radio">
     <h4>Radio Box</h4>
-    <ui-code language="html" :code="code.html"></ui-code>
-    <ui-code language="js" :code="code.js"></ui-code>
+    <ui-code language="html" :code="code1.html"></ui-code>
+    <ui-code language="js" :code="code1.js"></ui-code>
     <p>
       <ui-radio name="sex" value="M" :model="gender" @change="chooseSex">Male</ui-radio>
       <ui-radio name="sex" value="F" :model="gender" @change="chooseSex">Female</ui-radio>
@@ -15,8 +15,6 @@
 
 <script>
 import UiCode from '../components/code';
-import codeTemplate from '../usages/radio/template.html';
-import codeScript from '../usages/radio/script.html';
 
 export default {
   components: {
@@ -25,7 +23,10 @@ export default {
   data() {
     return {
       gender: '',
-      code: {}
+      code1: {
+        html: require('../snippets/radio/demo1-template.html'),
+        js: require('../snippets/radio/demo1-script.html')
+      }
     };
   },
   methods: {
@@ -36,10 +37,6 @@ export default {
       console.info('gender');
       console.log(this.gender);
     }
-  },
-  created() {
-    this.code.html = this.renderHTML(codeTemplate);
-    this.code.js = this.renderJS(codeScript);
   }
 };
 </script>