+++ CHANGELOG.md
@@ -17,6 +17,7 @@ __`<ui-table`>__
 ### Update docs
 
 - `<ui-table>`
+- `<ui-tabs>`
 
 ## 0.1.4 beta